Hpx is windows compat

+++ create-box-app.js
@@ -1,5 +1,6 @@
 const { exec } = require('child_process');
 const fs = require('fs');
+const path = require('path');
 const webpackContent = require('./resources/webpack.content');
 const htmlContent = require('./resources/indexhtml.content');
 const readmeContent = require('./resources/readme.content');
@@ -13,15 +14,16 @@ const { logGreen, logYellow, logLoader } = require('./resources/utils');
 
 module.exports = (args) => {
   const nameArg = args[0];
-  const path = process.env.PWD
   // Make new app folder with name.
-  const newAppPath = `${path}/${nameArg}`;
-  fs.mkdir(newAppPath, () => {
+  const newAppPath = path.join(process.cwd(), nameArg);
+  fs.mkdir(nameArg, (mkdirErr) => {
+    console.log(mkdirErr);
+    if (mkdirErr) throw new Error('Could not create project file, perhaps it already exists?');
     exec(`cd ${newAppPath}; npm init --yes; git init`, (err, stdout, stderr) => {
       logGreen('HyperBox: initialised npm ✅');
       logGreen('HyperBox: initialised git ✅');
       fs.writeFile(`${newAppPath}/package.json`, packageContent(nameArg), () => {})
-      logYellow('HyperBox: installing hyperbox-js...')
+      logYellow('HyperBox: installing hyperbox-js...');
       const stopLoader = logLoader();
       exec(`cd ${newAppPath}; npm install --save hyperbox-js`, () => {
         stopLoader();
@@ -34,12 +36,12 @@ module.exports = (args) => {
           fs.writeFile(
             gitIgnorePath, 
             gitignoreContent(nameArg),
-          () => { })
+          () => { });
           fs.mkdir(`${newAppPath}/public`, () => {
             logGreen('HyperBox: Added /public ✅');
             fs.copyFile(__dirname + '/resources/favicon.ico', `${newAppPath}/public/favicon.ico`, (...args) => {
               logGreen('HyperBox: Added /public/favicon.ico 📦 ✅');
-            })
+            });
           });
           fs.mkdir(`${newAppPath}/src`, () => {
             logGreen('HyperBox: Added /src ✅')
@@ -50,21 +52,23 @@ module.exports = (args) => {
               // Add the main box.
               fs.writeFile(`${newAppPath}/src/main/main.box.js`, mainContent(), () => {
                 logGreen('HyperBox: Added your first box 📦')
-              })
-            })
-          })
-          fs.writeFile(`${newAppPath}/server.js`, serverContent(), () => logGreen('HyperBox: Added server.js ✅'))
+              });
+            });
+          });
+          fs.writeFile(`${newAppPath}/server.js`, serverContent(), () => logGreen('HyperBox: Added server.js ✅'));
           fs.writeFile(`${newAppPath}/webpack.config.js`, webpackContent(nameArg), () => {
             logYellow('HyperBox: Installing dependencies...')
             const stopLoadingTwo = logLoader();
             exec(`cd ${newAppPath}; npm i`, () => {
-              stopLoadingTwo()
+              stopLoadingTwo();
               logGreen('HyperBox: Installed dependencies ✅ ⚡️')
-              fs.mkdir(`${newAppPath}/dist`, () => {});
+              fs.mkdir(`${newAppPath}/dist`, () => {
+                logGreen('HyperBox: Project setup complete!');
+              });
             });
-          })
+          });
       });
-      })
+      });
     });
   });
 }
